move request license button into own component
